chore: rename cost of living dataset to 2025
