+++ src/data/__tests__/cost-of-living-dataset.test.ts
@@ -1,7 +1,7 @@
-import { costOfLiving2024 } from '@/data/costOfLiving2024';
+import { costOfLiving2025 } from '@/data/costOfLiving2025';
 
 describe('cost-of-living dataset', () => {
   it('is current', () => {
-    expect(costOfLiving2024.baseYear).toBeGreaterThanOrEqual(new Date().getFullYear());
+    expect(costOfLiving2025.baseYear).toBeGreaterThanOrEqual(new Date().getFullYear());
   });
 });
