+++ src/data/costOfLiving2025.ts
@@ -13,7 +13,7 @@ export interface CostOfLivingDataset {
   regions: Record<string, RegionCost>;
 }
 
-export const costOfLiving2024: CostOfLivingDataset = {
+export const costOfLiving2025: CostOfLivingDataset = {
   baseYear: 2025,
   source: 'BEA Regional Price Parities 2024',
   regions: {
@@ -36,4 +36,4 @@ export const costOfLiving2024: CostOfLivingDataset = {
   },
 } as const;
 
-export type Region = keyof typeof costOfLiving2024.regions;
+export type Region = keyof typeof costOfLiving2025.regions;
